convert runtime to js
