--- app/static/_runtime
@@ -1,1 +1,0 @@
-../../init/template/base/static/_runtime
