+++ src/client/index.js
@@ -1,0 +1,3 @@
+export { HandlerRegistry } from "./handler-registry.js";
+export * from "./reactivity.js";
+export * from "./handlers.js";
