+++ src/client/utils.js
@@ -3,8 +3,9 @@ export const spaces_sep_by_comma = /\s*,\s*/;
 
 /**
  * Idempotent string conversion to kebab-case
+ * @param {string} str
  */
-export const toKebabCase = (str: string) => {
+export const toKebabCase = (str) => {
   let kebab = "";
   for (let index = 0; index < str.length; index++) {
     const char = str[index];
@@ -20,8 +21,9 @@ export const toKebabCase = (str: string) => {
 
 /**
  * Idempotent string conversion to PascalCase
+ * @param {string} str
  */
-export const toPascalCase = (str: string) => {
+export const toPascalCase = (str) => {
   let pascal = "";
   let toUpper = true;
